+++ src/js-framework/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.13.10
 
 - add `updateFinish` action after `click/fireEvent`
+- add `console.log/debug/info/warn/error`